Cap the number of conditions a query can contain
Every condition becomes a clause when a query is translated for a
search backend, so an unbounded query is expensive for consumers and
not just for us. Reject queries with more than MaxConditions (1000)
conditions in NewContactQuery, so the limit applies to programmatically
built queries as well as parsed ones.

Move the Simplify chain test to exercise Simplify directly, since
queries are now capped well below the size needed to show the
quadratic behaviour.
