+++ contactql/parse/parse_test.go
@@ -1,9 +1,9 @@
 package parse_test
 
 import (
+	"fmt"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/nyaruka/goflow/assets"
 	"github.com/nyaruka/goflow/assets/static"
@@ -537,29 +537,19 @@ func TestParseQueryTooComplex(t *testing.T) {
 	assert.Equal(t, contactql.ErrTooComplex, err.(*contactql.QueryError).Code())
 }
 
-// TestSimplifyLongChain checks that a long chain of same-op conditions is flattened correctly and, more
-// importantly, in linear time. Parsing `a or b or c ...` produces a left leaning tree and flattening it
-// used to copy the promoted children at every level, which is quadratic - a ~1.7MB query of this shape
-// took ~35s of CPU. The time bound here is very generous compared to the ~0.2s this now takes; it exists
-// only to catch a return of the quadratic behaviour.
-func TestSimplifyLongChain(t *testing.T) {
+func TestParseQueryTooManyConditions(t *testing.T) {
 	env := envs.NewBuilder().Build()
 	resolver := contactql.NewMockResolver(nil, nil, nil)
 
-	const n = 100000
-	query := "name=x" + strings.Repeat(" OR name=x", n)
-
-	start := time.Now()
-	parsed, err := parse.Query(env, query, resolver)
-	elapsed := time.Since(start)
-
+	// a query right on the limit is fine
+	ok := "name=x" + strings.Repeat(" OR name=x", contactql.MaxConditions-1)
+	parsed, err := parse.Query(env, ok, resolver)
 	require.NoError(t, err)
+	assert.Len(t, parsed.Root().(*contactql.BoolCombination).Children(), contactql.MaxConditions)
 
-	// all the conditions should have been promoted into a single OR node
-	root, ok := parsed.Root().(*contactql.BoolCombination)
-	require.True(t, ok, "root should be a bool combination")
-	assert.Equal(t, contactql.BoolOperatorOr, root.Operator())
-	assert.Len(t, root.Children(), n+1)
-
-	assert.Less(t, elapsed, 10*time.Second, "flattening a chain of %d conditions took %s - has it become quadratic again?", n, elapsed)
+	// one more is rejected
+	tooMany := "name=x" + strings.Repeat(" OR name=x", contactql.MaxConditions)
+	_, err = parse.Query(env, tooMany, resolver)
+	assert.EqualError(t, err, fmt.Sprintf("query contains more than %d conditions", contactql.MaxConditions))
+	assert.Equal(t, contactql.ErrTooComplex, err.(*contactql.QueryError).Code())
 }
